python/requirements.txt
matplotlib==3.5.3
numpy==1.23.2
opencv_python==4.6.0.66
pandas==1.4.4
Pillow==9.2.0
torch==1.12.1
torchvision==0.13.1
